move dynamic-agent.conf to /var/lib/nginx-agent

+++ scripts/packages/preinstall.sh
@@ -24,14 +24,19 @@ export AGENT_GROUP="${AGENT_GROUP:-$(id -ng)}"
 
 if [ "$ID" = "freebsd" ]; then
     AGENT_CONFIG_FILE=${AGENT_CONFIG_FILE:-"/usr/local/etc/nginx-agent/nginx-agent.conf"}
-    AGENT_DYNAMIC_CONFIG_DIR="/usr/local/etc/nginx-agent"
+    AGENT_DYNAMIC_CONFIG_DIR="/usr/local/var/lib/nginx-agent"
+    # Old location of agent-dynamic.conf 
+    OLD_DYNAMIC_CONFIG_DIR="/usr/local/etc/nginx-agent"
     mkdir -p /var/log/nginx-agent/
 else
-    AGENT_CONFIG_FILE=${AGENT_CONFIG_FILE:-"/etc/nginx-agent/nginx-agent.conf"}
-    AGENT_DYNAMIC_CONFIG_DIR="/etc/nginx-agent"
+    AGENT_CONFIG_FILE=${AGENT_CONFIG_FILE:-"/var/lib/nginx-agent/nginx-agent.conf"}
+    AGENT_DYNAMIC_CONFIG_DIR="/var/lib/nginx-agent"
+    # Old location of agent-dynamic.conf 
+    OLD_DYNAMIC_CONFIG_DIR="/etc/nginx-agent"
 fi
 
 AGENT_DYNAMIC_CONFIG_FILE="${AGENT_DYNAMIC_CONFIG_DIR}/agent-dynamic.conf"
+OLD_DYNAMIC_CONFIG_FILE="${OLD_DYNAMIC_CONFIG_DIR}/agent-dynamic.conf"
 AGENT_DYNAMIC_CONFIG_COMMENT="#
 # dynamic-agent.conf
 #
@@ -77,10 +82,17 @@ ensure_sudo() {
 load_config_values() {
     # If the file doesn't exist attempt to create it
     if [ ! -f "$AGENT_DYNAMIC_CONFIG_FILE" ]; then
-        printf "Could not find %s ... Creating file\n" ${AGENT_DYNAMIC_CONFIG_FILE}
-        mkdir -p ${AGENT_DYNAMIC_CONFIG_DIR}
-        printf "%s" "${AGENT_DYNAMIC_CONFIG_COMMENT}" | tee ${AGENT_DYNAMIC_CONFIG_FILE} > /dev/null
-        printf "Successfully created %s\n" "${AGENT_DYNAMIC_CONFIG_FILE}"
+
+        if [ -f "$OLD_DYNAMIC_CONFIG_FILE" ]; then 
+            mkdir -p ${AGENT_DYNAMIC_CONFIG_DIR}
+            ln "$OLD_DYNAMIC_CONFIG_FILE" "$AGENT_DYNAMIC_CONFIG_FILE"
+        else
+            printf "Could not find %s ... Creating file\n" ${AGENT_DYNAMIC_CONFIG_FILE}
+            mkdir -p ${AGENT_DYNAMIC_CONFIG_DIR}
+            printf "%s" "${AGENT_DYNAMIC_CONFIG_COMMENT}" | tee ${AGENT_DYNAMIC_CONFIG_FILE} > /dev/null
+            printf "Successfully created %s\n" "${AGENT_DYNAMIC_CONFIG_FILE}"
+        fi
+        
     fi
 
     # Check if there are existing values
